pyproject.toml
[project]
name = "elite-rag-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.29.0",
    "fastapi>=0.128.0",
    "pgvector>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sentence-transformers>=3.0.0",
    "sqlalchemy>=2.0.45",
    "strawberry-graphql[fastapi]>=0.289.0",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
